requirements.txt
streamlit 
plotly
FilesConnection
st_autorefresh
numpy
pandas
pydeck
